% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xSub_census_multiple_raw.R
\docType{data}
\name{xSub_census_multiple_raw}
\alias{xSub_census_multiple_raw}
\title{Census of multiple-source event-level datasets in xSub (updated June 15, 2020)}
\format{
A list with 4 elements:
\describe{
  \item{level0_bysource}{Countries organized by \code{data_source}. List object, where each sub-entry is also a list, containing entries for \code{data_source},\code{geo_window},\code{time_window},\code{dyad_type},\code{country_iso3},\code{country_name}.}
  \item{level0_bycountry}{Data sources organized by country. List of data.frames, where each row is a country, with columns for \code{country_iso3},\code{country_name},\code{geo_window},\code{time_window},\code{dyad_type},\code{data_sources}.}
  \item{level1}{Detailed information on data sources, countries and spatial levels of analysis. data.frame, where each row is a source-country combination, with columns for \code{data_source},\code{geo_window},\code{time_window},\code{dyad_type},\code{country_iso3},\code{country_name},\code{units}.}
  \item{all_countries}{Vector of all country ISO3 codes. Used by \code{get_xSub_multi}.}
}
}
\source{
\url{http://www.x-sub.org/}
}
\usage{
xSub_census_multiple_raw
}
\description{
A list of data sources, countries and levels of analysis available for download. Used by \code{info_xSub()}
}
\keyword{datasets}
